// Frontend/src/app/interfaces/Paciente.ts
export interface Paciente{
    id: number;
    tipoDocumento: string;
    numeroDocumento: number; 
    primerNombre: string; 
    segundoNombre: string;
    primerApellido: string; 
    segundoApellido: string;
    sexo: string; 
    edad: number; 
    telefono: number;
    correo: string;
    estadoPaciente: boolean;
}


    